add feature to drag cards
I added jQuery ui so I could use draggable and
sortable. I have to load the jQuery ui library 
at the bottom of the body. I'm not sure why the
gem isn't working correctly.

I also added some javascript that will trigger a 
timer when you click the button that moves a 
card to 'in progress'