fix jquery
For some reason I cannot remove the jquery
script from application.html.haml. It looks like
the gem isn't working properly. They suggest
removing query-rails, but I if do that it doesn't
work. 

It looks like there is a bug in a previous version
of jquery and the gem may not be updated to 
reflect this. One solution is to add 'clone' to
draggable, and then removing the element after
the move.