create controller and model for projects
I created a has_many/belongs_to relationship. At
this point I only store the project name. I would
have liked to not use the database for this but 
the search on github would have been more trouble
than it was worth.

I made the main landing page after you login the
projects#index, and moved the issues to the 
project#show page.